+++ src/duration.js
@@ -488,7 +488,7 @@ export default class Duration {
     // todo - this should keep the opts...
     const vals = this.toObject();
     normalizeValues(this.matrix, vals);
-    return Duration.fromObject(vals);
+    return clone(this, { values: vals }, true);
   }
 
   /**
@@ -552,7 +552,7 @@ export default class Duration {
       }
     }
 
-    return clone(this, { values: built }, true);
+    return clone(this, { values: built }, true).normalize();
   }
 
   /**