+++ src/interval.js
@@ -237,7 +237,10 @@ export default class Interval {
    * @return {[Interval]}
    */
   splitAt(...dateTimes) {
-    const sorted = dateTimes.map(friendlyDateTime).sort(),
+    const sorted = dateTimes
+        .map(friendlyDateTime)
+        .filter(d => this.contains(d))
+        .sort(),
       results = [];
     let { s } = this,
       i = 0;